refactors common code
